Rename factory to builder
